requirements.txt
tensorflow
numpy
Pillow
scikit-learn
matplotlib
seaborn
tqdm